+++ src/Util/EncryptionManager.php
@@ -39,7 +39,6 @@ class EncryptionManager
         } catch (\Exception $e) {
             \trigger_error($e->getMessage(), E_USER_ERROR);
         }
-        return false;
     }
 
     /**