+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.6]
+- Improve PHP 7.1 compatibility.
+- Better handling of HTTP errors.
+- Add new telemetry method.
+- Add tests and dev instructions.
+
 ## [2.0.5]
 - Fix sending of a call.